backend/pyproject.toml
[project]
name = "infer-forge-backend"
version = "0.1.0"
description = "INFER FORGE backend - automatizační platforma pro strojírenskou firmu Infer s.r.o."
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.0",
    "sqlalchemy[asyncio]==2.0.35",
    "asyncpg==0.30.0",
    "alembic==1.13.0",
    "celery[redis]==5.4.0",
    "pydantic==2.9.0",
    "pydantic-settings==2.5.0",
    "python-multipart==0.0.12",
    "httpx==0.27.0",
    "openpyxl==3.1.5",
    "pandas==2.2.0",
    "lxml==5.3.0",
    "aiosmtplib==3.0.0",
    "aioimaplib==1.1.0",
    "pytesseract==0.3.13",
    "pymupdf==1.24.0",
    "anthropic==0.37.0",
    "langchain==0.3.0",
    "sentence-transformers==3.1.0",
    "pgvector==0.3.0",
    "structlog==24.4.0",
    "sentry-sdk[fastapi]==2.14.0",
    "python-jose[cryptography]==3.3.0",
    "pytest==8.3.0",
    "pytest-asyncio==0.24.0",
    "pytest-cov==6.0.0",
    "aiosqlite==0.20.0",
    "ruff==0.6.0",
    "mypy==1.11.0",
    "email-validator>=2.3.0",
    "bcrypt>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "C4", "T20"]
ignore = []

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
